+++ gulpfile.js
@@ -6,7 +6,7 @@ var sourcemaps = require('gulp-sourcemaps');
 var perfectionist = require('perfectionist');
 
 var files = ['index.js'];
-var watchFiles = ['index.js', 'gulpfile.js', 'demo/src/**/**', 'test/src/**/**'];
+var watchFiles = ['index.js', 'gulpfile.js', 'examples/src/**/**', 'test/src/**/**'];
 
 gulp.task('lint', function () {
 	var eslint = require('gulp-eslint');
@@ -16,14 +16,14 @@ gulp.task('lint', function () {
 		.pipe(eslint.failAfterError());
 });
 
-gulp.task('demo', function () {
-	return gulp.src('demo/*.js', {read: false})
+gulp.task('examples', function () {
+	return gulp.src('examples/*.js', {read: false})
 		.pipe(mocha({timeout: 1000000}));
 });
 
 gulp.task('htmlcopy', function () {
-	return gulp.src(['./demo/src/html/index.html'], {base: './demo/src/html/'})
-		.pipe(gulp.dest('./demo/dist/html/'));
+	return gulp.src(['./examples/src/html/index.html'], {base: './examples/src/html/'})
+		.pipe(gulp.dest('./examples/dist/html/'));
 });
 
 gulp.task('test', function () {
@@ -32,25 +32,24 @@ gulp.task('test', function () {
 });
 
 gulp.task('css', function () {
-	return gulp.src('./demo/src/css/*.css')
+	return gulp.src('./examples/src/css/*.css')
 		.pipe(sourcemaps.init())
 		.pipe(postcss([lazysprite({
-			imagePath: './demo/src/slice',
-			stylesheetPath: './demo/dist/css',
-			spritePath: './demo/dist/sprites',
+			imagePath: './examples/src/slice',
+			stylesheetPath: './examples/dist/css',
+			spritePath: './examples/dist/sprites',
 			smartUpdate: true,
-			nameSpace: '',
-			logLevel: 'debug'  // 'debug','info','slient'
+			logLevel: 'debug'
 		}), perfectionist({
 			maxAtRuleLength: false
 		})]))
 		.pipe(sourcemaps.write("."))
-		.pipe(gulp.dest('./demo/dist/css'));
+		.pipe(gulp.dest('./examples/dist/css'));
 });
 
 gulp.task('default', ['htmlcopy', 'css', 'watch']);
 
 gulp.task('watch', function () {
-	// gulp.watch(watchFiles, ['css', 'demo', 'lint']);
+	// gulp.watch(watchFiles, ['css', 'test', 'lint']);
 	gulp.watch(watchFiles, ['css', 'lint']);
 });