+++ index.js
@@ -47,8 +47,8 @@ module.exports = postcss.plugin('postcss-lazysprite', function (options) {
 		padding: options.padding ? options.padding : 10,
 		nameSpace: options.nameSpace || '',
 		outputDimensions: options.outputDimensions || true,
-		smartUpdate: options.smartUpdate || true,
-		logLevel: options.logLevel || 'debug'  // 'debug','info','slient'
+		smartUpdate: options.smartUpdate || false,
+		logLevel: options.logLevel || 'info'  // 'debug','info','slient'
 	}, options);
 
 	// Option `imagePath` is required
@@ -391,7 +391,7 @@ function saveSprites(images, options, sprites) {
 
 				// If this file is up to date
 				if (sprite.isFromCache) {
-					log(options.logLevel, 'lv2', ['Lazysprite:', gutil.colors.yellow(sprite.path), 'unchanged.']);
+					log(options.logLevel, 'lv3', ['Lazysprite:', gutil.colors.yellow(sprite.path), 'unchanged.']);
 					deferred.resolve(sprite);
 					return deferred.promise;
 				}
@@ -411,7 +411,7 @@ function saveSprites(images, options, sprites) {
 				// Save new file version
 				return fs.writeFileAsync(sprite.path, new Buffer(sprite.image, 'binary'))
 					.then(function () {
-						log(options.logLevel, 'lv1', ['Lazysprite:', gutil.colors.green(sprite.path), 'generated.']);
+						log(options.logLevel, 'lv2', ['Lazysprite:', gutil.colors.green(sprite.path), 'generated.']);
 						return sprite;
 					});
 			})
@@ -623,10 +623,12 @@ function log(logLevel, level, content) {
 
 	switch (logLevel) {
 	case 'slient':
-		output = false;
+		if (level !== 'lv1') {
+			output = false;
+		}
 		break;
 	case 'info':
-		if (level === 'lv2') {
+		if (level === 'lv3') {
 			output = false;
 		}
 		break;